Validate Quantity products can buy

+++ TechShopSolution.WebApp/Controllers/CartController.cs
@@ -51,7 +51,7 @@ namespace TechShopSolution.WebApp.Controllers
                 var cartItem = new CartItemViewModel()
                 {
                     Id = id,
-                    Instock = (int)product.ResultObject.instock,
+                    Instock = product.ResultObject.instock,
                     Code = product.ResultObject.code,
                     Slug = product.ResultObject.slug,
                     Price = product.ResultObject.unit_price,
